refactor(gardener/classifiers): extract parseVerdictJson + validateAndGroundNodes to shared module
Pulls the verdict-JSON parser, verdict/severity validators, summary
clamping, and tree-path grounding helper out of anthropic.ts into
`verdict-parse.ts`. No behaviour change — `anthropic.ts` re-exports
`parseVerdictJson` and `validateAndGroundNodes` so existing test
imports keep resolving.

Prep for #269: the upcoming `claude-cli.ts` classifier needs the
same parsing + grounding pipeline. Landing the extraction as its own
PR per bingran's review guidance so the follow-up classifier PR is
purely additive and easy to review/revert.

Refs #269
